==> src/5_ESMPreprocStep2.Rmd <==
---
title: "ESM Preprocessing Step 2"
author: "DVAL Lab"
date: "`r Sys.Date()`"
output: html_document
---

# Overview

This R Markdown file follows Step 2 from the Revol et al. (2024) ESM preprocessing guide:  
[https://preprocess.esmtools.com](https://preprocess.esmtools.com)  
It focuses on describing the design and sampling scheme of the ESM data.

---

```{r setup, include=FALSE}
library(here)
library(dplyr)
library(esmtools)
library(ggplot2)
library(lubridate)
library(ggpubr)
library(ggTimeSeries)

# Define path to the Step 1 output
input_path <- here("output", "EMA_and_baseline1.rds")
```

#Read in data
```{r}
d <- readRDS(input_path)
```

#From Revol et al, (2024), "Ensuring the coherence of the timestamps and the order of the observations is crucial for the analysis of EMA data. We will focus on this critical task, with specific objectives including examining both key timestamp variables (e.g., init_time, completion_time, study_day) and the 'session_id' variable."
#To achieve this, we will create a subsetted dataframe with only ID#, timestamp variables and observation numbers are included.

```{r}
d1 <- d %>%
  dplyr::select(participant_id, init_time, completion_time, session_id, study_day, obsno)
```

#Coherence within observations of chronological sequence
```{r}
any(d1$init_time > d1$completion_time)
```

#Interpretation: Timestamp check within observations inidcates no issues.

#Coherence between observations
```{r}
d1 %>%
  arrange(participant_id, init_time) %>%  # Sort by participant and init_time
  group_by(participant_id) %>%
  mutate(
    init_lag_issue = lag(init_time) > init_time,
    completion_lag_issue = lag(completion_time) > completion_time,
    session_id_lag_issue = lag(session_id) > session_id,
    study_day_lag_issue = lag(study_day) > study_day
  ) %>%
  filter(init_lag_issue | completion_lag_issue | session_id_lag_issue | study_day_lag_issue)
```
#Interpretation 5/15/25: session_id was flagged as having greater lagged values. This may not be an issue, depending on how the survey was set up 

#Check observation order: step skipped because this was verified during the creation of obsno variable in previous RMD.

#Calendar plot: code below will not work for our data because we do not have a sent time
#We can gain useful information about the data by examining the frequency of surveys initiated by date.
```{r}
# Organize the dataset and compute the number of beep sent for each date
d1_date %>% d1
  mutate(date = as.Date(sent)) %>%
  group_by(completion_time) %>%
  summarise(n_sent = n()) # Number of observation sent for each date
ggplot_calendar_heatmap(d1, 'completion', 'n_sent') +
  scale_fill_continuous(low = 'green', high = 'red') +
  theme(
    legend.position = "bottom",
    axis.text.x = element_text(angle = 45, hjust = 1, size = 6),
    axis.text.y = element_text(size = 6),
    strip.text = element_text(size = 8)
  )
```

#Sampling scheme plot: A sampling scheme plot provides a broad view of all the beeps, in particular here the beeps sent to the participants. It is a useful tool to identify any missingness, patterns in missingness, or design-related issues. The beeps can be displayed regarding different timeframes (e.g., beep level, day level)
#Beep-level: below is a plot of missed beeps for all participants who missed a beep.
```{r}
d1 %>% filter(is.na(init_time)) %>%
  group_by(participant_id, obsno) %>%
  ggplot(aes(x = obsno, y = factor(participant_id))) +
  geom_point(size=1.5) +
  theme(axis.text.x = element_text(angle = 90))
```
#Duration plot: Step skipped: irregularities in duration of participant enrollment were checked in the Step 1: Import Data and Preliminary Preprocessing.
#The following steps were skipped because we do not have a 'time_sent' variable: Time sent beeps, Delay to send, Quantity of beeps sent, Delay between two sent beeps. Dyadic delay in sending as also skipped because we are not examining dyads.

#Save updates to data
```{r}
saveRDS(d, here("output", "EMA_and_baseline2.rds"))
```








#References

#Revol, J., Carlier, C., Lafit, G., Verhees, M., Sels, L., & Ceulemans, E. (2024). Preprocessing ESM data: a step-by-step framework, tutorial website, R package, and reporting templates.
